Read the cookies on the server to populate the store
