+++ package.json
@@ -56,6 +56,7 @@
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/svelte": "^3.1.0",
     "@testing-library/user-event": "^14.0.4",
+    "@types/cookie": "^0.5.1",
     "@types/dompurify": "^2.3.3",
     "@types/js-cookie": "^3.0.2",
     "@types/lodash-es": "^4.17.6",
@@ -107,6 +108,7 @@
     "@fontsource/raleway": "^4.5.5",
     "@supabase/supabase-js": "^1.33.2",
     "clsx": "^1.1.1",
+    "cookie": "^0.5.0",
     "dompurify": "^2.3.6",
     "js-cookie": "^3.0.1",
     "marked": "^4.0.12",
