+++ src/lib/features/listening-history/index.ts
@@ -1,1 +1,2 @@
-export { addToListeningHistory, listeningHistory, cookieName } from './model/store';
+export { addToListeningHistory, listeningHistory } from './model/store';
+export { populateListeningHistory } from './model/populate-listening-history';
